requirements.txt
certifi==2020.12.5
joblib==1.0.0
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.1
transformers==4.6.1
neptune-client
urlexpander==0.0.36
demoji==0.4.0